requirements.txt
absl-py==0.11.0
algorithmia==1.6.1
algorithmia-api-client==1.1.0
aniso8601==1.2.0
APScheduler==3.6.3
asn1crypto==1.4.0
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
blinker==1.4
boto3==1.14.51
botocore==1.17.51
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==2.1.4
Cython==0.29.21
docutils==0.15.2
enum34==1.1.10
Flask==0.12.1
Flask-APScheduler==1.11.0
flask-lambda==0.0.4
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.3.0
imutils==0.5.3
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python==4.1.0.25
opt-einsum==3.3.0
picamera==1.13
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.24.1
PyJWT==1.7.1
pylibdmtx==0.1.9
pylint==2.6.0
pyOpenSSL==17.0.0
pyshortcuts==1.7.1
pytesseract==0.3.7
python-dateutil==2.8.1
pyttsx3==2.90
pytz==2020.1
PyYAML==3.12
requests==2.24.0
requests-oauthlib==1.3.0
RPi.GPIO==0.7.0
rsa==4.6
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sounddevice==0.4.1
SoundFile==0.10.3.post1
SQLAlchemy==1.3.20
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow @ https://github.com/bitsy-ai/tensorflow-arm-bin/releases/download/v2.4.0-rc2/tensorflow-2.4.0rc2-cp37-none-linux_armv7l.whl
tensorflow-estimator==2.4.0
termcolor==1.1.0
tflite-runtime @ https://github.com/google-coral/pycoral/releases/download/release-frogfish/tflite_runtime-2.5.0-cp37-cp37m-linux_armv7l.whl
threadpoolctl==2.1.0
toml==0.10.1
twilio==6.46.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.10
Werkzeug==0.16.0
whichcraft==0.6.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.4.0
